Separate slice type to own file
